feat: Major refactoring. Deleted Hide.cs
Change the value type of the Scripture _text. From List<string> to
List<Word>.

Replace the behavior to hide words. Remove the Hide class and its
methods.

Implemented Word class to handle the behavior of each word in the
Scripture Text

Implemented new methods to handle the hidding process and validations in
the Scripture class